Use a single static dom::parser
@XrXr pointed out to me that `rb_raise` isn't quite safe in C++. Long story short
`rb_raise` bypass the C++ destructors which can lead to various leaks etc.

The good new is that `simdjson::dom::parser` is designed to be re-used.
When you call it again it re-use the memory it previously allocated,
so by instantiating a single parser once, and then using only that one,
we should avoid leaks.

Ref: https://github.com/anilmaurya/fast_jsonparser/pull/12