Be able to limit output to a number of entries
This is the same behavior as memory_profiler. We want to be able to
limit the number of entries on the output.

The only issue I see is one arguments limit the number of string to
limit + the number of location where the string was allocated. Maybe it
is a little bit misleading but the main goal is to limit size of the
output. So it is probably ok.