Allow to ship this gem with precompiled binaries
TL;DR I'd like to propose releasing this gem with precompiled
binaries built for different platforms and different ABI version
(fat gem).

I'm currently working on a tool to help the Ruby community ship gems
with precompiled binaries with the intent to make `bundle install`
much faster for everyone. The main bottleneck when installing gems
in a project is the compilation of native extensions.

[cibuildgem](https://github.com/Shopify/cibuildgem) modestly tries to
follow the same approach as what the python community did with
[cibuildwheel](https://cibuildwheel.pypa.io/en/stable/).
It works with a native compilation using CI runners (GitHub
it the only supported vendor for now) and tries to be as easy to setup
as possible.

The CI workflow in this commit was generated with the cibuildgem CLI
which reads the gemspec and determine what ruby versions needs to be
compiled and tested against.